Now using PropStoreNode instead of NodeData.

+++ TypeSafePropertyBag/ObjectIdDictionary/SimpleObjectIdDictionary.cs
@@ -3,7 +3,7 @@
 namespace DRM.TypeSafePropertyBag
 {
     using CompositeKeyType = UInt64;
-    using ObjectIdType = UInt32;
+    using ObjectIdType = UInt64;
     using PropIdType = UInt32;
     using PropNameType = String;
